Add scheduling to delete-invalid-orders

+++ src/server/graphql/nexus/mutations/scheduled-event.mutations/run-scheduled-event.mutation/delete-invalid-orders.field.ts
@@ -1,6 +1,8 @@
 import { extendType } from "@nexus/schema";
-import { OrderSide, OrderStatus, OrderType } from "@prisma/client";
+import { OrderEventType, OrderSide, OrderStatus, OrderType, Recurrence } from "@prisma/client";
 import { oneLine } from "common-tags";
+
+const DELETE_INVALID_ORDERS_ID = "OrderEventDeleteInvalidOrders";
 
 export const deleteInvalidOrders = extendType({
 	type: "RunScheduledEvent",
@@ -13,6 +15,39 @@ export const deleteInvalidOrders = extendType({
 			`,
 			authorize: (parent, args, { isEasyCron }) => isEasyCron(),
 			resolve: async ({ scheduledEvents }, args, { prisma }) => {
+				const event = await prisma.orderEvent.findOne({
+					where: { type: OrderEventType.DeleteInvalidOrders }
+				});
+
+				if (!event) {
+					/**
+					 * @description Create this scheduled event, if it does not currently exist.
+					 *     Use a daily recurrence, so that it only executes once per day (at the
+					 *     start of each day).
+					 */
+					await prisma.orderEvent.create({
+						data: {
+							type: OrderEventType.DeleteInvalidOrders,
+							scheduledEvent: {
+								create: {
+									id: DELETE_INVALID_ORDERS_ID,
+									recurrence: Recurrence.Daily,
+									hour: 0,
+									minute: 0
+								}
+							}
+						}
+					});
+				}
+
+				const didTrigger = scheduledEvents.some(
+					({ id }) => id === DELETE_INVALID_ORDERS_ID
+				);
+
+				if (!didTrigger) {
+					return 0;
+				}
+
 				const { count } = await prisma.order.deleteMany({
 					where: {
 						side: OrderSide.Buy,
